+++ TDTChocolate/TDTFoundationAdditions.h
@@ -3,6 +3,7 @@
 #import "FoundationAdditions/TDTSuppressPerformSelectorLeakWarning.h"
 #import "FoundationAdditions/NSArray+TDTFunctionalAdditions.h"
 #import "FoundationAdditions/NSSet+TDTFunctionalAdditions.h"
+#import "FoundationAdditions/NSDictionary+TDTFunctionalAdditions.h"
 #import "FoundationAdditions/TDTObjectOrDefault.h"
 #import "FoundationAdditions/NSArray+TDTAdditions.h"
 #import "FoundationAdditions/NSString+TDTAdditions.h"