+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 2.?.?, ??-???-2014
 
+  * Add dictionary method for transforming values.
   * Add method for removing characters from string.
 
 2.1.0, 30-May-2014